@@ -3,7 +3,7 @@ package com.funkyradish.funky_radish
 import android.app.Activity
 import android.content.Context
 import android.net.ConnectivityManager
-import android.preference.PreferenceManager
+import androidx.preference.PreferenceManager
 import android.util.Log
 import android.widget.Toast
 import com.android.volley.*
@@ -17,19 +17,23 @@ import io.realm.mongodb.Credentials
 import io.realm.mongodb.sync.SyncConfiguration
 import org.json.JSONException
 
-fun getToken(context: Context): String {
+fun getToken(context: Context): String? {
     val preferences = PreferenceManager.getDefaultSharedPreferences(context)
-    return preferences.getString(Constants.FR_TOKEN, "")
+    var str = preferences.getString(Constants.FR_TOKEN, "")
+    if (str == null) {
+        str = ""
+    }
+    return str
 }
 
-fun setToken(context: Context, token: String) {
+fun setToken(context: Context, token: String ) {
     val preferences = PreferenceManager.getDefaultSharedPreferences(context)
     val editor = preferences.edit()
     editor.putString(Constants.FR_TOKEN, token)
     editor.apply()
 }
 
-fun getUserEmail(context: Context): String {
+fun getUserEmail(context: Context): String? {
     val preferences = PreferenceManager.getDefaultSharedPreferences(context)
     return preferences.getString(Constants.FR_USER_EMAIL, "")
 }
@@ -63,21 +67,21 @@ fun toggleOfflineMode(context: Context) {
     editor.apply()
 }
 
-fun isConnected(context: Context): Boolean {
-    val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-    val networkInfo = connectivityManager.activeNetworkInfo
-    return networkInfo != null && networkInfo.isConnected
-}
+//fun isConnected(context: Context): Boolean {
+//    val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+//    val networkInfo = connectivityManager.activeNetworkInfo
+//    return networkInfo != null && networkInfo.isConnected
+//}
 
 //TODO: get rid of the importRecipes
-fun register(activity: Activity, queue: RequestQueue, email: String, password: String, importRecipes: List<Recipe?>, callback: (success: Boolean) -> Unit) {
+fun register(activity: Activity, queue: RequestQueue, email: String, password: String, callback: (success: Boolean) -> Unit) {
 
     // Structure user data
-    val json = JSONObject().apply({
+    val json = JSONObject().apply {
         put("email", email)
         put("password", password)
         put("admin", false)
-    })
+    }
 
     // User request
     val userRequest = object : JsonObjectRequest(Method.POST, Constants.ENDPOINT, json,
@@ -107,8 +111,8 @@ fun register(activity: Activity, queue: RequestQueue, email: String, password: S
                 val response = error.networkResponse
                 if (error is ServerError && response != null) {
                     try {
-                        val json = String(response.data)
-                        val userErrorResponse = GsonBuilder().create().fromJson(json, UserErrorResponse::class.java)
+                        val jsonResponse = String(response.data)
+                        val userErrorResponse = GsonBuilder().create().fromJson(jsonResponse, UserErrorResponse::class.java)
                         errorMessage = userErrorResponse.message
                     } catch (e2: JSONException) {
                         e2.printStackTrace()
@@ -130,55 +134,57 @@ fun register(activity: Activity, queue: RequestQueue, email: String, password: S
         }
     }
 
-    if(isConnected(activity.applicationContext)) {
-        queue.add(userRequest)
-    }
-    else {
-        Toast.makeText(
-                activity.applicationContext,
-                "Unable to connect to internet",
-                Toast.LENGTH_SHORT).show()
-    }
+    queue.add(userRequest)
+
+//    if(isConnected(activity.applicationContext)) {
+//        queue.add(userRequest)
+//    }
+//    else {
+//        Toast.makeText(
+//                activity.applicationContext,
+//                "Unable to connect to internet",
+//                Toast.LENGTH_SHORT).show()
+//    }
 
 }
 
 fun login(activity: Activity, queue: RequestQueue, email: String, password: String, importRecipes: List<Recipe?>, callback: (success: Boolean) -> Unit) {
 
     // Structure user data
-    val json = JSONObject().apply({
+    val json = JSONObject().apply {
         put("email", email)
         put("password", password)
-    })
+    }
 
     val authRequest = object : JsonObjectRequest(Method.POST, Constants.AuthEndpoint, json,
-            Response.Listener<JSONObject> { response ->
-                val body = response.toString()
-                val response = GsonBuilder().create().fromJson(body, UserResponse::class.java)
+            Response.Listener<JSONObject> { resp ->
+                val body = resp.toString()
+                val authResponse = GsonBuilder().create().fromJson(body, UserResponse::class.java)
 
-                if (response.message == "Enjoy your token, ya filthy animal!") {
-                    setToken(activity.applicationContext, response.token)
-                    createRealmUser(response.token, importRecipes, callback, activity)
+                if (authResponse.message == "Enjoy your token, ya filthy animal!") {
+                    setToken(activity.applicationContext, authResponse.token)
+                    createRealmUser(authResponse.token, importRecipes, callback, activity)
                 }
                 else {
                     Toast.makeText(
                             activity.applicationContext,
-                            response.message,
+                            authResponse.message,
                             Toast.LENGTH_SHORT).show()
 
                     callback(false)
                 }
             },
             Response.ErrorListener { error ->
-                Log.d("API", "Authorization error: ${error}")
+                Log.d("API", "Authorization error: $error")
                 error.printStackTrace()
 
                 var errorMessage = error.toString()
 
-                val response = error.networkResponse
-                if (error is ServerError && response != null) {
+                val authResponse = error.networkResponse
+                if (error is ServerError && authResponse != null) {
                     try {
-                        val json = String(response.data)
-                        val userErrorResponse = GsonBuilder().create().fromJson(json, UserErrorResponse::class.java)
+                        val jsonResp = String(authResponse.data)
+                        val userErrorResponse = GsonBuilder().create().fromJson(jsonResp, UserErrorResponse::class.java)
                         errorMessage = userErrorResponse.message
                     } catch (e2: JSONException) {
                         e2.printStackTrace()
@@ -200,16 +206,18 @@ fun login(activity: Activity, queue: RequestQueue, email: String, password: Stri
         }
     }
 
-    //TODO: Do we really need to check connection?
-    if(isConnected(activity.applicationContext)) {
-        queue.add(authRequest)
-    }
-    else {
-        Toast.makeText(
-                activity.applicationContext,
-                "Unable to connect to internet",
-                Toast.LENGTH_SHORT).show()
-    }
+    queue.add(authRequest)
+
+//    //TODO: Do we really need to check connection?
+//    if(isConnected(activity.applicationContext)) {
+//        queue.add(authRequest)
+//    }
+//    else {
+//        Toast.makeText(
+//                activity.applicationContext,
+//                "Unable to connect to internet",
+//                Toast.LENGTH_SHORT).show()
+//    }
 
 }
 
